Use the parameter of the image handle to allocate memory.
The original code will use Vulkan.Image.Handle to query the requirement of
the image to be allocated memory, which is unreasonable, since the handle
for the image has been given as the parameter "image" already. When reuse
this function to create other images, the program would crash in the
driver side for this reason.